+++ anvil_consortium_manager/tests/utils.py
@@ -3,8 +3,11 @@ from unittest import mock
 import google.auth.credentials
 import google.auth.transport.requests
 import responses
+from faker import Faker
 
 from ..anvil_api import AnVILAPIClient
+
+fake = Faker()
 
 
 class AnVILAPIMockTestMixin:
@@ -33,6 +36,11 @@ class AnVILAPIMockTestMixin:
         self.anvil_response_mock.start()
         # Get an instance of the API client to access entry points?
         self.api_client = AnVILAPIClient()
+        # Set the auth session service account email here, since some functions need it.
+        self.service_account_email = fake.email()
+        AnVILAPIClient().auth_session.credentials.service_account_email = (
+            self.service_account_email
+        )
 
     def tearDown(self):
         super().tearDown()